lagt till try/catch-satser på ALLT och några IF-kontroller.

+++ Infrastructure/Services/AdressService.cs
@@ -2,6 +2,7 @@
 
 using Infrastructure.Entities;
 using Infrastructure.Repositories;
+using System.Diagnostics;
 
 namespace Infrastructure.Services;
 
@@ -17,39 +18,96 @@ public class AdressService
 
     public AdressEntity CreateAdress(string streetName, string postalCode, string city)
     {
-        var adressEntity = _adressRepository.Get(x => x.StreetName == streetName && x.PostalCode == postalCode && x.City == city);
-        adressEntity ??= _adressRepository.Create(new AdressEntity { StreetName = streetName, PostalCode = postalCode, City = city });
+        try
+        {
+            var adressEntity = _adressRepository.Get(x => x.StreetName == streetName && x.PostalCode == postalCode && x.City == city);
+            adressEntity ??= _adressRepository.Create(new AdressEntity { StreetName = streetName, PostalCode = postalCode, City = city });
 
-        return adressEntity;
+            if (adressEntity != null)
+            {
+                
+                return adressEntity;
+            }
+        }
+        catch (Exception ex) { Debug.WriteLine("ERROR :: " + ex.Message); }
+        return null!;
+
+
 
     }
 
     public AdressEntity GetAdress(string streetName, string postalCode, string city)
     {
-        var adressEntity = _adressRepository.Get(x => x.StreetName == streetName && x.PostalCode == postalCode && x.City == city);
-        return adressEntity;
+        try
+        {
+            var adressEntity = _adressRepository.Get(x => x.StreetName == streetName && x.PostalCode == postalCode && x.City == city);
+            if (adressEntity != null)
+            {
+                return adressEntity;
+            }
+        }
+        catch (Exception ex) { Debug.WriteLine("ERROR :: " + ex.Message); }
+        return null!;
+
     }
 
     public AdressEntity GetAdressById(int id)
     {
-        var adressEntity = _adressRepository.Get(x => x.Id == id);
-        return adressEntity;
+        try
+        {
+            var adressEntity = _adressRepository.Get(x => x.Id == id);
+            if (adressEntity != null)
+            {
+                return adressEntity;
+            }
+        }
+        catch (Exception ex) { Debug.WriteLine("ERROR :: " + ex.Message); }
+        return null!;
+
     }
 
     public IEnumerable<AdressEntity> GetAdresses()
     {
-        var adresses = _adressRepository.GetAll();
-        return adresses;
+        try
+        {
+            var adresses = _adressRepository.GetAll();
+            if (adresses != null)
+            {
+                return adresses;
+            }
+        }
+        catch (Exception ex) { Debug.WriteLine("ERROR :: " + ex.Message); }
+        return null!;
+
     }
 
     public AdressEntity UpdateAdress(AdressEntity adressEntity)
     {
-        var updatedAdressEntity = _adressRepository.Update(x => x.Id == adressEntity.Id, adressEntity);
-        return updatedAdressEntity;
+        try
+        {
+            var updatedAdressEntity = _adressRepository.Update(x => x.Id == adressEntity.Id, adressEntity);
+            if (updatedAdressEntity != null)
+            {
+                return updatedAdressEntity;
+            }
+        }
+        catch (Exception ex) { Debug.WriteLine("ERROR :: " + ex.Message); }
+        return null!;
+
     }
 
-    public void DeleteAdress(int id)
+    public bool DeleteAdress(int id)
     {
-        _adressRepository.Delete(x => x.Id == id);
+        try
+        {
+            if (id != 0)
+            {
+                _adressRepository.Delete(x => x.Id == id);
+                return true;
+            }
+              
+        }
+        catch (Exception ex) { Debug.WriteLine("ERROR :: " + ex.Message); }
+        return false;
     }
 }
