+++ Infrastructure/Services/LeagueService.cs
@@ -2,6 +2,7 @@
 
 using Infrastructure.Entities;
 using Infrastructure.Repositories;
+using System.Diagnostics;
 
 namespace Infrastructure.Services;
 
@@ -16,41 +17,91 @@ public class LeagueService
 
     public LeagueEntity CreateLeague(string leagueName, string nation)
     {
-        var leagueEntity = _leagueRepository.Get(x => x.LeagueName == leagueName && x.Nation == nation);
-        leagueEntity ??= _leagueRepository.Create(new LeagueEntity { LeagueName = leagueName, Nation = nation });
+        try
+        {
+            var leagueEntity = _leagueRepository.Get(x => x.LeagueName == leagueName && x.Nation == nation);
+            leagueEntity ??= _leagueRepository.Create(new LeagueEntity { LeagueName = leagueName, Nation = nation });
+            if (leagueEntity != null)
+            {
+                return leagueEntity;
+            }
+        }
+        catch (Exception ex) { Debug.WriteLine("ERROR :: " + ex.Message); }
+        return null!;
 
-        return leagueEntity;
 
     }
 
     public LeagueEntity GetLeague(string leagueName, string nation)
     {
-        var leagueEntity = _leagueRepository.Get(x => x.LeagueName == leagueName && x.Nation == nation);
-        return leagueEntity;
+        try
+        {
+            var leagueEntity = _leagueRepository.Get(x => x.LeagueName == leagueName && x.Nation == nation);
+            if (leagueEntity != null)
+            {
+                return leagueEntity;
+            }
+        }
+        catch (Exception ex) { Debug.WriteLine("ERROR :: " + ex.Message); }
+        return null!;
     }
 
 
 
     public LeagueEntity GetLeagueById(int id)
     {
-        var leagueEntity = _leagueRepository.Get(x => x.Id == id);
-        return leagueEntity;
+        try
+        {
+            var leagueEntity = _leagueRepository.Get(x => x.Id == id);
+            if (leagueEntity != null)
+            {
+                return leagueEntity;
+            }
+        }
+        catch (Exception ex) { Debug.WriteLine("ERROR :: " + ex.Message); }
+        return null!;
+
     }
 
     public IEnumerable<LeagueEntity> GetLeagues()
     {
-        var leagues = _leagueRepository.GetAll();
-        return leagues;
+        try
+        {
+            var leagues = _leagueRepository.GetAll();
+            if (leagues != null)
+            {
+                return leagues;
+            }
+        }
+        catch (Exception ex) { Debug.WriteLine("ERROR :: " + ex.Message); }
+        return null!;
     }
 
     public LeagueEntity UpdateLeague(LeagueEntity leagueEntity)
     {
-        var updatedLeagueEntity = _leagueRepository.Update(x => x.Id == leagueEntity.Id, leagueEntity);
-        return updatedLeagueEntity;
+        try
+        {
+            var updatedLeagueEntity = _leagueRepository.Update(x => x.Id == leagueEntity.Id, leagueEntity);
+            if (updatedLeagueEntity != null)
+            {
+                return updatedLeagueEntity;
+            }
+        }
+        catch (Exception ex) { Debug.WriteLine("ERROR :: " + ex.Message); }
+        return null!;
     }
 
-    public void DeleteLeague(int id)
+    public bool DeleteLeague(int id)
     {
-        _leagueRepository.Delete(x => x.Id == id);
+        try
+        {
+            if (id != 0)
+            {
+                _leagueRepository.Delete(x => x.Id == id);
+                return true;
+            }
+        }
+        catch (Exception ex) { Debug.WriteLine("ERROR :: " + ex.Message); }
+        return false;
     }
 }
